Clean up Cream Colors config structure
